+++ meteor/server/api/playout/__tests__/lib.test.ts
@@ -67,28 +67,28 @@ describe('selectNextPart', () => {
 	test('from nothing', () => {
 		{
 			// default
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 0, part: defaultParts[0] })
 		}
 
 		{
 			// first isnt playable
 			defaultParts[0].playable = false
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 1, part: defaultParts[1] })
 		}
 
 		{
 			// nextSegmentId is set
 			defaultPlaylist.nextSegmentId = segment3
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 6, part: defaultParts[6], consumesNextSegmentId: true })
 		}
 
 		{
 			// nextSegmentId is set (and first there isnt playable)
 			defaultPlaylist.nextSegmentId = segment2
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 4, part: defaultParts[4], consumesNextSegmentId: true })
 		}
 	})
@@ -96,28 +96,28 @@ describe('selectNextPart', () => {
 	test('from nothing - allow unplayable', () => {
 		{
 			// default
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 0, part: defaultParts[0] })
 		}
 
 		{
 			// first isnt playable
 			defaultParts[0].playable = false
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 0, part: defaultParts[0] })
 		}
 
 		{
 			// nextSegmentId is set
 			defaultPlaylist.nextSegmentId = segment3
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 6, part: defaultParts[6], consumesNextSegmentId: true })
 		}
 
 		{
 			// nextSegmentId is set (and first there isnt playable)
 			defaultPlaylist.nextSegmentId = segment2
-			const nextPart = selectNextPart(defaultPlaylist, null, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, null, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 3, part: defaultParts[3], consumesNextSegmentId: true })
 		}
 	})
@@ -126,14 +126,14 @@ describe('selectNextPart', () => {
 		const previousPartInstance = defaultParts[4].toPartInstance()
 		{
 			// default
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 5, part: defaultParts[5] })
 		}
 
 		{
 			// next isnt playable
 			defaultParts[5].playable = false
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 6, part: defaultParts[6] })
 		}
 
@@ -141,14 +141,14 @@ describe('selectNextPart', () => {
 			// nextSegmentId is set
 			defaultParts[0].playable = false
 			defaultPlaylist.nextSegmentId = segment1
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 1, part: defaultParts[1], consumesNextSegmentId: true })
 		}
 
 		{
 			// nextSegmentId is set (and first there isnt playable)
 			defaultPlaylist.nextSegmentId = segment2
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 4, part: defaultParts[4], consumesNextSegmentId: true })
 		}
 	})
@@ -160,21 +160,21 @@ describe('selectNextPart', () => {
 
 		{
 			// single part is orphaned
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 4, part: defaultParts[4] })
 		}
 
 		{
 			// whole segment is orphaned/deleted
 			defaultParts = defaultParts.filter((p) => p.segmentId !== previousPartInstance.segmentId)
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 3, part: defaultParts[3] })
 		}
 
 		{
 			// no parts after
 			defaultParts = defaultParts.filter((p) => p.segmentId !== segment3)
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual(null)
 		}
 
@@ -182,7 +182,7 @@ describe('selectNextPart', () => {
 			// no parts after, but looping
 			defaultPlaylist.loop = true
 			defaultParts = defaultParts.filter((p) => p.segmentId !== segment3)
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts())
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts())
 			expect(nextPart).toEqual({ index: 0, part: defaultParts[0] })
 		}
 	})
@@ -191,14 +191,14 @@ describe('selectNextPart', () => {
 		const previousPartInstance = defaultParts[4].toPartInstance()
 		{
 			// default
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 5, part: defaultParts[5] })
 		}
 
 		{
 			// next isnt playable
 			defaultParts[5].playable = false
-			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, getSegmentsAndParts(), false)
+			const nextPart = selectNextPart(defaultPlaylist, previousPartInstance, null, getSegmentsAndParts(), false)
 			expect(nextPart).toEqual({ index: 5, part: defaultParts[5] })
 		}
 	})